[IFX] - testcode  실패 상황 보충. (#43)

+++ src/main/java/buravel/buravel/modules/plan/PlanController.java
@@ -30,6 +30,7 @@ public class PlanController {
     private final PlanService planService;
     private final PlanValidator planValidator;
     private final PatchPlanValidator patchPlanValidator;
+    private final PlanRepository planRepository;
 
     @PostMapping
     public ResponseEntity createPlan(@RequestBody @Valid PlanDto planDto, @CurrentUser Account account, Errors errors) {
@@ -58,6 +59,9 @@ public class PlanController {
 
     @GetMapping("/{id}")
     public ResponseEntity getPlan(@PathVariable Long id) throws NotFoundException {
+        if (!planRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         EntityModel<PlanWithPostResponseDto> result = planService.getPlanWithPlanId(id);
         return ResponseEntity.ok(result);
     }
@@ -69,7 +73,6 @@ public class PlanController {
             EntityModel<Errors> jsr303 = ErrorResource.of(errors);
             return ResponseEntity.badRequest().body(jsr303);
         }
-
         // 여행 날짜 request 체크
         patchPlanValidator.validate(patchplanRequestDto,errors);
         if (errors.hasErrors()) {
@@ -87,7 +90,10 @@ public class PlanController {
     }
 
     @DeleteMapping("/{planId}")
-    public ResponseEntity deletePlan(@PathVariable Long planId, @CurrentUser Account account) throws NotFoundException  {
+    public ResponseEntity deletePlan(@PathVariable Long planId, @CurrentUser Account account) throws NotFoundException {
+        if (!planRepository.existsById(planId)) {
+            return ResponseEntity.notFound().build();
+        }
         planService.deletePlan(planId, account);
         return ResponseEntity.ok().build();
     }
